Remove virt_wifi's REGULATORY_WIPHY_SELF_MANAGED
This turns out not to be necessary for the default VirtWifi wireless
network. It causes a WARN_ON and a netlink failure message in
NL80211_CMD_GET_REG because it claims to manage the regulatory domain
itslef, and then doesn't actually manage it.

The problem is fixed by not claiming a regulatory domain.

Bug: 119771992
Test: Ran build making the NL80211_CMD_GET_REG call, observed no warning.
Change-Id: Ic8ceec40d9e50da908da228a69e3cd976f0c0312
Signed-off-by: A. Cody Schuffelen <schuffelen@google.com>